+++ c/benjamin_model/src/cell.h
@@ -35,7 +35,7 @@ typedef struct {
 } Cell;
 
 Cell* initCell(int x, int y, int lx, int ly, double dr,
-	       double incell, int seed);
+	       double incell, long seed);
 void deleteCell(Cell* cell);
 void initFieldSquare(Cell* cell, int x0, int y0, int dx, int dy, double phi0);
 void initField(Cell* cell, double** field);
